pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "prc25"
version = "0.1.0"
description = "An open machine learning model to infer the fuel consumed of a flown flight."
readme = "README.md"
license = "GPL-3.0-or-later"
requires-python = ">=3.10"
dependencies = [
    "flash-linear-attention>=0.3.2",
    "polars>=1.34",
    "rich>=14",
    "tomli>=2 ; python_full_version < '3.11'",
]

[project.optional-dependencies]
cli = ["typer>=0.19"]
plots = ["matplotlib>=3.10", "networkx[default]>=3.4.2"]

[dependency-groups]
docs = ["mkdocs-material[imaging]>=9.6.14", "mkdocstrings-python>=1.16.11"]
wandb = ["wandb>=0.22.2"]
wrangler = ["ipykernel>=6.30.1", "pandas>=2.3.3"]

[tool.ruff]
include = ["notebooks/**/*.ipynb", "scripts/**/*.py", "src"]
line-length = 100

[tool.ruff.lint]
select = [
    "E",
    "W",      # pycodestyle
    "F",      # pyflakes
    "I",      # isort
    "NPY",    # numpy
    "NPY201", # numpy
    "DTZ",    # flake8-datetimez
    "RUF",
]
ignore = [
    "F722", # see astral-sh/ruff#9860
]
